+++ ddv/src/main/java/community/ddv/domain/user/service/UserService.java
@@ -5,23 +5,23 @@ import community.ddv.domain.board.repository.CommentRepository;
 import community.ddv.domain.board.repository.ReviewRepository;
 import community.ddv.domain.certification.Certification;
 import community.ddv.domain.certification.CertificationRepository;
+import community.ddv.domain.certification.constant.CertificationStatus;
 import community.ddv.domain.certification.constant.RejectionReason;
-import community.ddv.domain.user.dto.UserDTO.AccountDeleteDto;
-import community.ddv.domain.user.dto.UserDTO.LoginResponseDto;
-import community.ddv.domain.user.dto.UserDTO.NicknameUpdateRequestDto;
-import community.ddv.domain.user.dto.UserDTO.LoginDto;
-import community.ddv.domain.user.dto.UserDTO.NicknameUpdateResponseDto;
-import community.ddv.domain.user.dto.UserDTO.OneLineIntroRequestDto;
-import community.ddv.domain.user.dto.UserDTO.OneLineIntroResponseDto;
-import community.ddv.domain.user.dto.UserDTO.PasswordUpdateDto;
-import community.ddv.domain.user.dto.UserDTO.SignUpDto;
-import community.ddv.domain.user.dto.UserDTO.TokenDto;
-import community.ddv.domain.user.dto.UserDTO.UserInfoResponseDto;
+import community.ddv.domain.user.constant.Role;
+import community.ddv.domain.user.dto.SignDto.AccountDeleteDto;
+import community.ddv.domain.user.dto.SignDto.LoginDto;
+import community.ddv.domain.user.dto.SignDto.LoginResponseDto;
+import community.ddv.domain.user.dto.SignDto.SignUpDto;
+import community.ddv.domain.user.dto.SignDto.TokenDto;
+import community.ddv.domain.user.dto.UserInfoDto.NicknameUpdateRequestDto;
+import community.ddv.domain.user.dto.UserInfoDto.NicknameUpdateResponseDto;
+import community.ddv.domain.user.dto.UserInfoDto.OneLineIntroRequestDto;
+import community.ddv.domain.user.dto.UserInfoDto.OneLineIntroResponseDto;
+import community.ddv.domain.user.dto.UserInfoDto.PasswordUpdateDto;
+import community.ddv.domain.user.dto.UserInfoDto.UserInfoResponseDto;
 import community.ddv.domain.user.entity.User;
 import community.ddv.domain.user.repository.UserRepository;
 import community.ddv.global.component.JwtProvider;
-import community.ddv.domain.certification.constant.CertificationStatus;
-import community.ddv.domain.user.constant.Role;
 import community.ddv.global.exception.DeepdiviewException;
 import community.ddv.global.exception.ErrorCode;
 import java.time.DayOfWeek;
